Struct anemo_build::manual::MethodBuilder

pub struct MethodBuilder { /* private fields */ }

Method builder.

This builder can be used to manually define an RPC method, which can be added to an RPC service.

Example

let say_hello_method = Method::builder()
    .name("say_hello")
    .route_name("SayHello")
    // Provide the path to the Request type
    .request_type("crate::common::HelloRequest")
    // Provide the path to the Response type
    .response_type("crate::common::HelloResponse")
    // Provide the path to the Codec to use
    .codec_path("crate::common::JsonCodec")
    .build();

Implementations

impl MethodBuilder

pub fn name(self, name: impl AsRef<str>) -> Self

Set the name for this Method.

This value will be used for generating the client functions for calling this Method.

Generally this is formatted in snake_case.

pub fn route_name(self, route_name: impl AsRef<str>) -> Self

Set the route_name for this Method.

This value will be used as part of the route for calling this method. Routes have the form: /<package_name>.<service_name>/<method_route_name>

Generally this is formatted in PascalCase.

pub fn comment(self, comment: impl AsRef<str>) -> Self

Add a comment string that should be included as a doc comment for this Method.

pub fn request_type(self, request_type: impl AsRef<str>) -> Self

Set the path to the Rust type that should be use for the Request type of this method.

pub fn response_type(self, response_type: impl AsRef<str>) -> Self

Set the path to the Rust type that should be use for the Response type of this method.

pub fn codec_path(self, codec_path: impl AsRef<str>) -> Self

Set the path to the Rust type that should be used as the Codec for this method.

Currently the codegen assumes that this type implements Default.

pub fn server_handler_return_raw_bytes(self, use_raw_bytes: bool) -> Self

Set whether or not the server handler should use raw bytes for the response.

pub fn build(self) -> Method

Build a Method

Panics if name, route_name, request_type, response_type, or codec_path weren’t set.